Create sampled vars as inputs before V3Sched
This fixes t_clocking_inout.pl -fno-gate
