+++ src/Verilator.cpp
@@ -81,6 +81,7 @@
 #include "V3ProtectLib.h"
 #include "V3Randomize.h"
 #include "V3Reloop.h"
+#include "V3Sampled.h"
 #include "V3Sched.h"
 #include "V3Scope.h"
 #include "V3Scoreboard.h"
@@ -436,6 +437,9 @@ static void process() {
             // Differs from V3Active, because identical clocks may be pushed
             // down to a module and now be identical
             V3ActiveTop::activeTopAll(v3Global.rootp());
+
+            // Remove SAMPLED
+            V3Sampled::sampledAll(v3Global.rootp());
 
             if (v3Global.opt.stats()) V3Stats::statsStageAll(v3Global.rootp(), "PreOrder");
